refactor(crowdfundings): Rename lib/scheduler to lib/jobs
